requirements.txt
ffmpeg-python
openai-whisper
pyyaml
torch
